requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangochannelsrestframework==1.2.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
jiter==0.5.0
msgpack==1.0.8
oauthlib==3.2.2
openai==0.28.0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyOpenSSL==24.2.1
python3-openid==3.2.0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==74.0.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==13.0.1
zope.interface==7.0.3
